requirements.txt
torch
flask
flask-login
flask-sqlalchemy
werkzeug
transformers
python-dotenv
flask-wtf
wtforms
torch
